chore(rig-sqlite): move inline test modules to sibling files
